fix arm64: asm_and_imm sign-extends imm via uint32_t instead of int64_t
(uint64_t)(uint32_t)(-16) = 0x00000000FFFFFFF0, but 64-bit AND needs
0xFFFFFFFFFFFFFFF0. Also asm_and_imm with negative immediates fails
logic_imm encoding because the truncated value isnt a valid bitmask
pattern. Fixed by using (uint64_t)(int64_t)imm.

Affects VLA alignment (and r,r,#-16) and other negative AND masks.

+++ src/codegen_asm.h
@@ -2017,7 +2017,7 @@ static void asm_mfence(SecBuf *s) {
 static void asm_and_imm(SecBuf *s, VReg r, int size, int32_t imm) {
 #ifdef ARCH_ARM64
     int sf = (size == 8) ? 1 : 0;
-    arm64_and_imm(s, sf, REG(r), REG(r), (uint64_t)(uint32_t)imm);
+    arm64_and_imm(s, sf, REG(r), REG(r), (uint64_t)(int64_t)imm);
 #else
     x86_and_ri(s, size, REG(r), imm);
 #endif
